+++ docs/CHANGELOG.md
@@ -5,7 +5,7 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [[v1.0.5-SNAPSHOT] - NOT-YET-RELEASED](/tas/alfresco-tas-tester/commits/v1.0.5)
+## [[v1.0.5] - 2016-12-13](/tas/alfresco-tas-tester/commits/v1.0.5)
 ### Added
 - ability to remove site from favorites
 - remove ErrorModel from model package (moved to restapi project)